backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database and ORM
supabase==2.3.4
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.29.0

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7

# AI and ML
openai==1.10.0
anthropic==0.8.1
langchain==0.1.4
langchain-openai==0.0.5

# MCP (Model Context Protocol)
# Installed separately via: npx @modelcontextprotocol/server-playwright

# Redis for caching
redis==5.0.1
hiredis==2.3.2

# Excel parsing
openpyxl==3.1.2
pandas==2.1.4
xlrd==2.0.1

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Logging and monitoring
loguru==0.7.2
sentry-sdk==1.39.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
